refactor: Reorganize card creation and display views into modular CardBuilder and CardView components with dedicated helpers.
