+++ src/views/CardBuilder/CardBuilderHeader.svelte
@@ -1,0 +1,34 @@
+<header class="maker__header">
+  <p class="maker__eyebrow">Travel Cards</p>
+  <h1 class="maker__title">Card Builder</h1>
+  <p class="maker__subtitle">
+    Select the languages you speak and choose the items you need to communicate
+    quickly while traveling.
+  </p>
+</header>
+
+<style>
+  .maker__header {
+    display: grid;
+    gap: 0.75rem;
+  }
+
+  .maker__eyebrow {
+    margin: 0;
+    font-size: 0.85rem;
+    letter-spacing: 0.18em;
+    text-transform: uppercase;
+    color: #7a6a50;
+  }
+
+  .maker__title {
+    margin: 0;
+    font-size: clamp(2rem, 4vw, 3.25rem);
+  }
+
+  .maker__subtitle {
+    margin: 0;
+    max-width: 38rem;
+    color: #4f463a;
+  }
+</style>
